Add function to replace event object enrichment in apis
